% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{odb_get_taxons}
\alias{odb_get_taxons}
\alias{odb_import_taxons}
\alias{odb_test}
\alias{odb_get_jobs}
\alias{odb_get_log}
\alias{odb_get_affected_ids}
\alias{odb_get_locations}
\alias{odb_import_locations}
\alias{odb_get_people}
\alias{odb_import_people}
\alias{odb_get_plants}
\alias{odb_import_plants}
\alias{odb_import_measurements}
\title{Base functions}
\usage{
odb_get_taxons(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_import_taxons(data, odb_cfg = odb_config(), common = list())

odb_test(odb_cfg = odb_config())

odb_get_jobs(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_get_log(job_id, odb_cfg = odb_config())

odb_get_affected_ids(job_id, odb_cfg = odb_config())

odb_get_locations(params = list(), odb_cfg = odb_config(),
  simplify = TRUE)

odb_import_locations(data, odb_cfg = odb_config(), common = list())

odb_get_people(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_import_people(data, odb_cfg = odb_config(), common = list())

odb_get_plants(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_import_plants(data, odb_cfg = odb_config(), common = list())

odb_import_measurements(data, odb_cfg = odb_config(), common = list())
}
\arguments{
\item{params}{named list, named vector or character string. Represents the parameters to be sent for the server, such as "valid=TRUE" to return only valid taxons. See \code{\link{odb_params}}}

\item{odb_cfg}{list. A configuration object, as generated by \code{\link{odb_config}}}

\item{simplify}{logical. Should the results be 'unlisted' before being returned?}

\item{data}{The data to be imported. The format is quite flexible, but normally a named \code{data.frame} is the best choice.}

\item{common}{Optional. A named list containing fields which are common to all imported data.}

\item{job_id}{numeric. A single job id}
}
\value{
odb_get methods usually return a \code{data.frame}, but that depends on the JSON response received from the server.
}
\description{
These are some basic getter functions to interact with the API. See the full help on the package vignettes.
}
\examples{
cfg = odb_config()
# Test request
odb_test()
# Gets some taxons
param = list(level=210, valid=TRUE, limit=5)
taxons = odb_get_taxons(param, cfg)
print(taxons)
\dontrun{
data = data.frame( 
  name = c("Euterpe oleracea", "Euterpe edulis"), 
  parent = "Euterpe"
)
common = list(level = "species")
odb_import_taxons(data, cfg, common)
}
}
