% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{odb_test}
\alias{odb_test}
\alias{odb_get_bibreferences}
\alias{odb_get_datasets}
\alias{odb_get_biocollections}
\alias{odb_get_individuals}
\alias{odb_get_languages}
\alias{odb_get_locations}
\alias{odb_get_measurements}
\alias{odb_get_persons}
\alias{odb_get_projects}
\alias{odb_get_taxons}
\alias{odb_get_traits}
\alias{odb_get_vouchers}
\title{Connection and data download functions}
\usage{
odb_test(odb_cfg = odb_config())

odb_get_bibreferences(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_get_datasets(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_get_biocollections(
  params = list(),
  odb_cfg = odb_config(),
  simplify = TRUE
)

odb_get_individuals(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_get_languages(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_get_locations(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_get_measurements(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_get_persons(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_get_projects(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_get_taxons(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_get_traits(params = list(), odb_cfg = odb_config(), simplify = TRUE)

odb_get_vouchers(params = list(), odb_cfg = odb_config(), simplify = TRUE)
}
\arguments{
\item{odb_cfg}{list. A configuration object, as generated by \code{\link{odb_config}}}

\item{params}{named list, named vector or character string. Represents the parameters to be sent for the server, such as "valid=1" to return only valid taxons. See \code{\link{odb_params}}}

\item{simplify}{logical. Should the results be 'unlisted' before being returned?}
}
\value{
odb_get methods usually return a \code{data.frame} is simplify=TRUE, but that depends on the JSON response received from the server.
}
\description{
These are some basic functions to interact with the API and download data. 
See the full help on the package vignettes.
}
\examples{
\dontrun{
cfg = odb_config()
# Test request
odb_test()
# Gets some taxons
param = list(level=210, valid=TRUE, limit=5)
taxons = odb_get_taxons(param, cfg)
print(taxons)
}
}
