% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{odb_taxonLevelCodes}
\alias{odb_taxonLevelCodes}
\alias{odb_traitTypeCodes}
\alias{odb_nomenclaturalTypes}
\alias{odb_detModifiers}
\title{OpenDataBio Import Helpers}
\usage{
odb_taxonLevelCodes(level = "all")

odb_traitTypeCodes()

odb_nomenclaturalTypes()

odb_detModifiers()
}
\arguments{
\item{level}{character. Taxon string level, 'species', 'fam.'. Default 'all'}
}
\value{
data.frame or vector depending on function
}
\description{
Functions to retrieve codes and ids required to import data. Taxon levels, Trait types, Nomenclatural Types
}
\examples{
odb_taxonLevelCodes()
odb_taxonLevelCodes(level='clade')
odb_traitTypeCodes()
odb_nomenclaturalTypes()
odb_detModifiers()
}
