% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{odb_config}
\alias{odb_config}
\title{Generate the connection configuration}
\usage{
odb_config(base_url, token, api_version, ...)
}
\arguments{
\item{base_url}{character. The base URL for the API calls. It should start with the protocol ("http") 
and end with "api". Defaults to OpenDataBio test server.}

\item{token}{character. The API token for a registered user. The data that can be accessed will vary
from user to user! Leave blank for anonymous login.}

\item{api_version}{character. Use this to specify the API version to connect to. You can set this to NULL to connect to the latest API, but this might break your code.}

\item{\dots}{Further parameters to be passed to add_headers.}
}
\value{
list
}
\description{
Use this function to generate a base configuration for all the other
functions.
}
\examples{
cfg = odb_config(
   base_url = "http://localhost/opendatabio/api", 
   token = "ABCDEF", 
   api_version = "v0"
)
}
