% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{odb_get_taxons}
\alias{odb_get_taxons}
\title{Base functions}
\usage{
odb_get_taxons(params = c(), odb_cfg = odb_config())
}
\arguments{
\item{params}{named list, named vector or character string. Represents the parameters to be sent for the server, such as "valid=TRUE" to return only valid taxons. See \code{\link{odb_params}}}

\item{odb_cfg}{list. A configuration object, as generated by \code{\link{odb_config}}}
}
\value{
usually, a \code{data.frame}, but that depends on the JSON response received from the server.
}
\description{
These are some basic getter functions to show the uses of the API
}
\examples{
cfg = odb_config("http://localhost/opendatabio/api")
param = list(level=210, valid=TRUE, external = FALSE)
taxons = odb_get_taxons(param, cfg)
}
