% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{odb_params}
\alias{odb_params}
\title{Convert query parameters}
\usage{
odb_params(params = list())
}
\arguments{
\item{params}{named list or character string.}
}
\value{
character
}
\description{
A convenience helper to transform R named lists to HTML syntax.
}
\details{
This is normally used inside the getter functions. Note that if this function
receives an unnamed character string, it returns the same string. This can be used to construct more
complex queries.
}
\examples{
odb_params(list(level=210, valid=TRUE))
odb_params("search=edulis")
}
