% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendatabio.r
\docType{package}
\name{opendatabio}
\alias{opendatabio}
\title{OpenDataBio: Imports and Exports Data From an OpenDataBio Server}
\description{
This package provides functions for downloading and uploading data from an OpendataBio repository.
}
