% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_locations}
\alias{plot_locations}
\title{Simple plotting function}
\usage{
plot_locations(locations, ...)
}
\arguments{
\item{locations}{data.frame, as returned by \code{\link{odb_get_locations}}}

\item{\dots}{additional graphic parameters}
}
\description{
This is a plotting example for using rgeos with OpenDataBio data.
}
