% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sp_to_df}
\alias{sp_to_df}
\title{Converts SpatialPolygonsDataFrame data for handling}
\usage{
sp_to_df(data, name_arg = NA, adm_level_arg = NA)
}
\arguments{
\item{data}{SpatialPolygonsDataFrame to be converted}

\item{name_arg}{a vector of names to be used}

\item{adm_level_arg}{a vector of adm_levels to be used}
}
\description{
This function converts SpatialPolygonsDataFrame to data.frame for easier importing to OpenDataBio.
}
\details{
This function requires that the suggested package 'rgeos' is installed. It fills some of the fields using 
the data supplied, and works best with http://gadm.org data. NOTE: this function is deprecated and will be replaced by a more in-depth vignette.
}
