% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wait_for_job}
\alias{wait_for_job}
\title{Wait for a running job}
\usage{
wait_for_job(job, odb_cfg = odb_config(), verbose = TRUE, interval = 10)
}
\arguments{
\item{job}{Either a number representing a job id, or the response from a odb_import_* call}

\item{odb_cfg}{list. A configuration object, as generated by \code{\link{odb_config}}}

\item{verbose}{logical. Should this function print job progress to screen?}

\item{interval}{numeric, time in seconds between prints}
}
\value{
data.frame, the final job object is silently returned to the user
}
\description{
This function sleeps until a job finishes. It can be used to "queue" 
jobs on the user side.
}
\examples{
\dontrun{
 wait_for_job(odb_import_locations(data, cfg), cfg)
}
}
