% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{odb_get_taxons}
\alias{odb_get_taxons}
\title{Base functions}
\usage{
odb_get_taxons(params = NULL)
}
\description{
These are some basic getter functions to show the uses of the API
}
\examples{
taxons = odb_get_taxons("level=210&valid=0&external=1")
}
